add decoder builder
