+++ src/parser.rs
@@ -1,10 +1,22 @@
 use charset::Charset;
+use std::{convert::TryFrom, result};
 
 use crate::lexer::{Token, Tokens};
 
-use std::convert::TryFrom;
+#[derive(thiserror::Error, Debug, Clone)]
+pub enum Error {
+    #[error("Unknown charset: {}", .0)]
+    UnknownCharset(String),
+    #[error("Unknown encoding: {}. Encoding can be only either 'Q' or 'B'.", .0)]
+    UnknownEncoding(char),
+    #[error("The encoded word is too big")]
+    EncodedWordTooBig,
+    #[error("Encoding is empty")]
+    EmptyEncoding,
+}
 
-pub type Result<T> = std::result::Result<T, Error>;
+type Result<T> = result::Result<T, Error>;
+
 pub type ClearText = Vec<u8>;
 pub type ParsedEncodedWords = Vec<ParsedEncodedWord>;
 
@@ -67,45 +79,34 @@ impl ParsedEncodedWord {
     }
 }
 
-#[derive(thiserror::Error, Debug, Clone)]
-pub enum Error {
-    #[error("Unknown charset: {}", .0)]
-    UnknownCharset(String),
-
-    #[error("Unknown encoding: {}. Encoding can be only either 'Q' or 'B'.", .0)]
-    UnknownEncoding(char),
-
-    #[error("The encoded word is too big")]
-    EncodedWordTooBig,
-
-    #[error("Encoding is empty")]
-    EmptyEncoding,
-}
-
 pub fn run(tokens: Tokens) -> Result<ParsedEncodedWords> {
     let parsed_encoded_words = convert_tokens_to_encoded_words(tokens)?;
     Ok(parsed_encoded_words)
 }
 
 fn convert_tokens_to_encoded_words(tokens: Tokens) -> Result<ParsedEncodedWords> {
-    tokens.into_iter().map(|token: Token| match token {
-        Token::ClearText(clear_text) => Ok(ParsedEncodedWord::ClearText(clear_text)),
-        Token::EncodedWord {
-            charset,
-            encoding,
-            encoded_text,
-        } => ParsedEncodedWord::convert_encoded_word(charset, encoding, encoded_text),
-    }).collect()
+    tokens
+        .into_iter()
+        .map(|token: Token| match token {
+            Token::ClearText(clear_text) => Ok(ParsedEncodedWord::ClearText(clear_text)),
+            Token::EncodedWord {
+                charset,
+                encoding,
+                encoded_text,
+            } => ParsedEncodedWord::convert_encoded_word(charset, encoding, encoded_text),
+        })
+        .collect()
 }
 
 #[cfg(test)]
 mod tests {
-
     use charset::Charset;
 
-    use crate::lexer;
-    use crate::parser;
-    use crate::parser::{Encoding, ParsedEncodedWord};
+    use crate::{
+        lexer,
+        parser::{self, Encoding, ParsedEncodedWord},
+        Decoder,
+    };
 
     /// Example taken from:
     /// https://datatracker.ietf.org/doc/html/rfc2047#section-8
@@ -114,7 +115,7 @@ mod tests {
     #[test]
     fn test_parse1() {
         let message = "=?US-ASCII?Q?Keith_Moore?=".as_bytes();
-        let tokens = lexer::run(&message).unwrap();
+        let tokens = lexer::run(&message, Decoder::new()).unwrap();
         let parsed = parser::run(tokens).unwrap();
 
         let expected = vec![ParsedEncodedWord::EncodedWord {
@@ -133,7 +134,7 @@ mod tests {
     #[test]
     fn test_parse2() {
         let message = "=?ISO-8859-1?Q?Keld_J=F8rn_Simonsen?=".as_bytes();
-        let tokens = lexer::run(&message).unwrap();
+        let tokens = lexer::run(&message, Decoder::new()).unwrap();
         let parsed = parser::run(tokens).unwrap();
 
         let expected = vec![ParsedEncodedWord::EncodedWord {
@@ -152,7 +153,7 @@ mod tests {
     #[test]
     fn test_parse3() {
         let message = "=?ISO-8859-1?Q?Andr=E9?=".as_bytes();
-        let tokens = lexer::run(&message).unwrap();
+        let tokens = lexer::run(&message, Decoder::new()).unwrap();
         let parsed = parser::run(tokens).unwrap();
 
         let expected = vec![ParsedEncodedWord::EncodedWord {
@@ -171,7 +172,7 @@ mod tests {
     #[test]
     fn test_parse4() {
         let message = "=?ISO-8859-1?B?SWYgeW91IGNhbiByZWFkIHRoaXMgeW8=?=".as_bytes();
-        let tokens = lexer::run(&message).unwrap();
+        let tokens = lexer::run(&message, Decoder::new()).unwrap();
         let parsed = parser::run(tokens).unwrap();
 
         let expected = vec![ParsedEncodedWord::EncodedWord {
